+++ spark/core/src/test/java/zingg/common/core/executor/TestObvDupeFilter.java
@@ -25,6 +25,7 @@ import zingg.common.client.ZFrame;
 import zingg.common.client.ZinggClientException;
 import zingg.common.client.pipe.Pipe;
 import zingg.common.client.util.ColName;
+import zingg.common.core.obviousdupes.ObvDupeFilter;
 import zingg.spark.client.SparkFrame;
 import zingg.spark.client.ZSparkSession;
 import zingg.spark.client.pipe.SparkPipe;
